% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_quality_analysis.R
\name{calculate_aqi}
\alias{calculate_aqi}
\title{Calculate Air Quality Index (AQI)}
\usage{
calculate_aqi(pm25, pm10, o3, no2)
}
\arguments{
\item{pm25}{PM2.5 measurements (the tiny particles)}

\item{pm10}{PM10 measurements (the bigger particles)}

\item{o3}{Ozone measurements}

\item{no2}{Nitrogen dioxide measurements}
}
\value{
A data frame with AQI scores and what they mean
}
\description{
This is my first attempt at calculating AQI! I used the EPA's method
but made it simpler to understand. It's not perfect, but it works!
}
