% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_impact.R
\name{calculate_daily_risk}
\alias{calculate_daily_risk}
\title{Calculate Daily Health Risk}
\usage{
calculate_daily_risk(aqi_data, days = 7)
}
\arguments{
\item{aqi_data}{A data frame with AQI values over time}

\item{days}{Number of days to look at (default is 7)}
}
\value{
A data frame with daily risk scores
}
\description{
This function looks at air quality over time and tries to guess
if it might make people sick. It's not perfect, but it's a start!
}
