% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health_impact.R
\name{calculate_health_impact}
\alias{calculate_health_impact}
\title{Calculate Health Impact Score}
\usage{
calculate_health_impact(aqi)
}
\arguments{
\item{aqi}{Air Quality Index values}
}
\value{
A data frame with health impact scores and warnings
}
\description{
This function tries to figure out how bad the air quality is for your health.
I used some basic rules from the WHO, but it's pretty simple!
}
