% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_quality_analysis.R
\name{calculate_no2_index}
\alias{calculate_no2_index}
\title{Calculate NO2 score}
\usage{
calculate_no2_index(no2)
}
\arguments{
\item{no2}{Nitrogen dioxide measurements}
}
\value{
NO2 score (0-300)
}
\description{
Calculate NO2 score
}
\keyword{internal}
