% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_quality_analysis.R
\name{calculate_o3_index}
\alias{calculate_o3_index}
\title{Calculate O3 score}
\usage{
calculate_o3_index(o3)
}
\arguments{
\item{o3}{Ozone measurements}
}
\value{
O3 score (0-300)
}
\description{
Calculate O3 score
}
\keyword{internal}
