% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/air_quality_analysis.R
\name{calculate_pm10_index}
\alias{calculate_pm10_index}
\title{Calculate PM10 score}
\usage{
calculate_pm10_index(pm10)
}
\arguments{
\item{pm10}{PM10 measurements}
}
\value{
PM10 score (0-300)
}
\description{
Calculate PM10 score
}
\keyword{internal}
