% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_aqi_trend}
\alias{plot_aqi_trend}
\title{Plot Air Quality Trends}
\usage{
plot_aqi_trend(aqi_data, pollutant = "AQI")
}
\arguments{
\item{aqi_data}{A data frame with dates and AQI values}

\item{pollutant}{Which pollutant to plot (default is "AQI")}
}
\value{
A ggplot object showing the trend
}
\description{
This function makes a graph of how air quality changes over time.
I used ggplot2 because it makes nice-looking plots!
}
