% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_daily_risk}
\alias{plot_daily_risk}
\title{Plot Daily Risk}
\usage{
plot_daily_risk(risk_data)
}
\arguments{
\item{risk_data}{A data frame with daily risk scores}
}
\value{
A ggplot object showing the daily risk
}
\description{
This function makes a line plot showing how risky the air quality
is over time. The higher the line, the worse it is!
}
